REFACTOR: api route and tests to discord-actions
