+++ controllers/discordactions.js
@@ -2,7 +2,9 @@ const { INTERNAL_SERVER_ERROR } = require("../constants/errorMessages");
 const admin = require("firebase-admin");
 const config = require("config");
 const jwt = require("jsonwebtoken");
+const userQuery = require("../models/users");
 const discordRolesModel = require("../models/discordactions");
+const { setUserDiscordNickname } = require("../services/discordService");
 
 /**
  * Creates a role
@@ -169,10 +171,45 @@ const setRoleIdleToIdleUsers = async (req, res) => {
   }
 };
 
+/**
+ * Patch Update user nicknames on discord
+ *
+ * @param req {Object} - Express request object
+ * @param res {Object} - Express response object
+ */
+
+const updateDiscordNicknames = async (req, res) => {
+  let response;
+  try {
+    const { dev } = req.query;
+    if (dev === "true") {
+      const usersInDiscord = await userQuery.getDiscordUsers();
+      const nonSuperUsers = usersInDiscord.filter((user) => !user.roles.super_user === true);
+      const batchUpdate = nonSuperUsers.map(async (user) => {
+        const { discordId, username } = user;
+        return await setUserDiscordNickname(username, discordId);
+      });
+      response = await Promise.all(batchUpdate);
+      return res.json({
+        numberOfUsersEffected: response.length,
+        message: `Users Nicknames updated successfully`,
+      });
+    } else {
+      return res.status(404).json({
+        message: "Users Nicknames not updated",
+      });
+    }
+  } catch (error) {
+    logger.error(`Error while updating nicknames: ${error}`);
+    return res.boom.badImplementation(INTERNAL_SERVER_ERROR);
+  }
+};
+
 module.exports = {
   createGroupRole,
   getAllGroupRoles,
   addGroupRoleToMember,
   updateDiscordImageForVerification,
   setRoleIdleToIdleUsers,
+  updateDiscordNicknames,
 };
