+++ test/integration/discordactions.test.js
@@ -95,13 +95,14 @@ describe("Discord actions", function () {
     let newGroupData;
     let allIds = [];
     before(async function () {
-      const addUsersPromises = userData.map((user) => userModel.add({ ...user }));
-      const responses = await Promise.all(addUsersPromises);
-      allIds = responses.map((response) => response.id);
+      const addUsersPromises = userData.map((user) => userModel.add({ ...user })); // Mapping over the mock users data and adding them to the usermodel(firestore instance)
+      const responses = await Promise.all(addUsersPromises); // waiting for all of them to get added to firestore instance and get an array of the responses after adding
+      allIds = responses.map((response) => response.id); // now getting Id's of the newly added users to the firestore instance
       newGroupData = groupData.map((group, index) => {
+        // adding createdBy peoperty to each group-object
         return {
           ...group,
-          createdBy: allIds[Math.min(index, allIds.length - 1)],
+          createdBy: allIds[Math.min(index, allIds.length - 1)], // It prevents accessing an index that doesn't exist, which could lead to errors.
         };
       });
 
@@ -163,6 +164,90 @@ describe("Discord actions", function () {
             expect(group).to.include.all.keys(expectedProps);
           });
           expect(res.body.message).to.equal("Roles fetched successfully!");
+          return done();
+        });
+    });
+  });
+
+  describe("POST /discord-actions/nicknames/sync", function () {
+    let superUserId;
+    let superUserAuthToken;
+    let fetchStub;
+    beforeEach(async function () {
+      fetchStub = sinon.stub(global, "fetch");
+      superUserId = await addUser(superUser);
+      superUserAuthToken = authService.generateAuthToken({ userId: superUserId });
+      const archivedRole = { roles: { archived: false } };
+      // const addUsersWithRolePromises = [
+      await addUser({ ...userData[0], archivedRole });
+      await addUser({ ...userData[1], archivedRole });
+      await addUser({ ...userData[2], archivedRole });
+      // ];
+      // await Promise.all(addUsersWithRolePromises);
+    });
+
+    afterEach(async function () {
+      sinon.restore();
+      await cleanDb();
+    });
+
+    it("should successfully update discord nicknames", function (done) {
+      fetchStub.returns(
+        Promise.resolve({
+          status: 200,
+          json: () => Promise.resolve(),
+        })
+      );
+      chai
+        .request(app)
+        .post(`/users/discord/nicknames?dev=true`)
+        .set("Cookie", `${cookieName}=${superUserAuthToken}`)
+        .end((err, res) => {
+          if (err) {
+            return done(err);
+          }
+          expect(res).to.have.status(200);
+          expect(res.body.message).to.be.equal("Users Nicknames updated successfully");
+          expect(res.body.numberOfUsersEffected).to.be.equal(3);
+          return done();
+        });
+    });
+    it("updates the nickname with username", function (done) {
+      // fetchStub.returns(
+      //   Promise.resolve({
+      //     status: 200,
+      //     json: () => Promise.resolve(updatedNicknameResponse),
+      //   })
+      // );
+      chai
+        .request(app)
+        .post(`/users/discord/nicknames?dev=true`)
+        .set("Cookie", `${cookieName}=${superUserAuthToken}`)
+        .end((err, res) => {
+          if (err) {
+            return done(err);
+          }
+          // expect(res).to.have.status(200);
+          // expect(res.body.message).to.be.equal("Users Nicknames updated successfully");
+          // expect(res.body.userAffected).to.be.equal("test-name-007");
+          return done();
+        });
+    });
+
+    it("returns an error for unsuccessful updating nicknames with POST method", function (done) {
+      fetchStub.returns(Promise.reject(new Error("User not verified")));
+
+      chai
+        .request(app)
+        .post(`/users/discord/nicknames?dev=true`)
+        .set("Cookie", `${cookieName}=${superUserAuthToken}`)
+        .end((err, res) => {
+          if (err) {
+            return done(err);
+          }
+          expect(res).to.have.status(500);
+          const response = res.body;
+          expect(response.message).to.be.equal("An internal server error occurred");
           return done();
         });
     });
