+++ test/integration/users.test.js
@@ -2003,52 +2003,95 @@ describe("Users", function () {
     });
   });
 
-  describe("POST /discord/nicknames", function () {
-    beforeEach(async function () {
-      fetchStub = Sinon.stub(global, "fetch");
-      userId = await addUser(userData[0]);
-    });
-    afterEach(async function () {
-      await cleanDb();
-      Sinon.restore();
-    });
-    it("returns 200 for successfully updating nicknames with POST method", function (done) {
-      fetchStub.returns(
-        Promise.resolve({
-          status: 200,
-          json: () => Promise.resolve(updatedNicknameResponse),
-        })
-      );
-      chai
-        .request(app)
-        .post(`/users/discord/nicknames?dev=true`)
-        .set("Cookie", `${cookieName}=${superUserAuthToken}`)
-        .end((err, res) => {
-          if (err) {
-            return done(err);
-          }
-          expect(res).to.have.status(200);
-          expect(res.body.message).to.be.equal("Users Nicknames updated successfully");
-          return done();
-        });
-    });
+  // describe("POST /discord/nicknames", function () {
+  //   // beforeEach(async function () {
+  //   //   fetchStub = Sinon.stub(global, "fetch");
+  //   //   userId = await addUser(userData[0]);
+  //   // });
+  //   // afterEach(async function () {
+  //   //   await cleanDb();
+  //   //   Sinon.restore();
+  //   // });
 
-    it("returns an error for unsuccessful updating nicknames with POST method", function (done) {
-      fetchStub.returns(Promise.reject(new Error("User not verified")));
+  //   let userId1;
+  //   let userId2;
+  //   let userId3;
 
-      chai
-        .request(app)
-        .post(`/users/discord/nicknames?dev=true`)
-        .set("Cookie", `${cookieName}=${superUserAuthToken}`)
-        .end((err, res) => {
-          if (err) {
-            return done(err);
-          }
-          expect(res).to.have.status(500);
-          const response = res.body;
-          expect(response.message).to.be.equal("An internal server error occurred");
-          return done();
-        });
-    });
-  });
+  //   beforeEach(async function () {
+  //     const rolesToBeAdded = {
+  //       archived: false,
+  //       in_discord: false,
+  //     };
+  //     userId1 = await addUser({ ...userData[0], roles: rolesToBeAdded });
+  //     userId2 = await addUser({ ...userData[1], roles: rolesToBeAdded });
+  //     userId3 = await addUser({ ...userData[2], roles: rolesToBeAdded });
+  //   });
+
+  //   afterEach(async function () {
+  //     await cleanDb();
+  //     Sinon.restore();
+  //   });
+
+  //   it("returns 200 for successfully updating nicknames with POST method", function (done) {
+  //     // fetchStub.returns(
+  //     //   Promise.resolve({
+  //     //     status: 200,
+  //     //     json: () => Promise.resolve(updatedNicknameResponse),
+  //     //   })
+  //     // );
+  //     chai
+  //       .request(app)
+  //       .post(`/users/discord/nicknames?dev=true`)
+  //       .set("Cookie", `${cookieName}=${superUserAuthToken}`)
+  //       .end((err, res) => {
+  //         if (err) {
+  //           return done(err);
+  //         }
+  //         console.log("response***", res);
+  //         expect(res).to.have.status(200);
+  //         expect(res.body.message).to.be.equal("Users Nicknames updated successfully");
+  //         return done();
+  //       });
+  //   });
+  //   it("updates the nickname with username", function (done) {
+  //     // fetchStub.returns(
+  //     //   Promise.resolve({
+  //     //     status: 200,
+  //     //     json: () => Promise.resolve(updatedNicknameResponse),
+  //     //   })
+  //     // );
+  //     chai
+  //       .request(app)
+  //       .post(`/users/discord/nicknames?dev=true`)
+  //       .set("Cookie", `${cookieName}=${superUserAuthToken}`)
+  //       .end((err, res) => {
+  //         if (err) {
+  //           return done(err);
+  //         }
+  //         console.log("response", res);
+  //         // expect(res).to.have.status(200);
+  //         // expect(res.body.message).to.be.equal("Users Nicknames updated successfully");
+  //         // expect(res.body.userAffected).to.be.equal("test-name-007");
+  //         return done();
+  //       });
+  //   });
+
+  //   it("returns an error for unsuccessful updating nicknames with POST method", function (done) {
+  //     fetchStub.returns(Promise.reject(new Error("User not verified")));
+
+  //     chai
+  //       .request(app)
+  //       .post(`/users/discord/nicknames?dev=true`)
+  //       .set("Cookie", `${cookieName}=${superUserAuthToken}`)
+  //       .end((err, res) => {
+  //         if (err) {
+  //           return done(err);
+  //         }
+  //         expect(res).to.have.status(500);
+  //         const response = res.body;
+  //         expect(response.message).to.be.equal("An internal server error occurred");
+  //         return done();
+  //       });
+  //   });
+  // });
 });
